switched to app router and fixed several warninggs
