+++ package.json
@@ -6,13 +6,12 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint ."
   },
   "dependencies": {
-    "eslint": "^8.57.1",
-    "eslint-config-next": "13.2.4",
+    "eslint-config-next": "^16.2.6",
     "framer-motion": "^10.6.0",
-    "next": "^13.5.7",
+    "next": "^16.2.6",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
@@ -21,12 +20,9 @@
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.7",
     "autoprefixer": "^10.4.19",
-    "eslint": "^8.36.0",
+    "eslint": "^9.39.0",
     "postcss": "^8.4.38",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.8.3"
-  },
-  "overrides": {
-    "ajv": "^8.17.1"
   }
 }
