+++ src/components/Education.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import React, { useRef } from "react"
 import { motion, useScroll } from "framer-motion"
 import LiIcon from "./LiIcon"
