+++ src/components/NavBar.tsx
@@ -1,7 +1,9 @@
+"use client";
+
 import Link from "next/link";
 import React from "react";
 import Logo from "./Logo";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import {
   DribbbleIcon,
   GithubIcon,
@@ -11,8 +13,14 @@ import {
 } from "./Icons";
 import { motion } from "framer-motion";
 
-const CustomLink = ({ href, title, className = " " }) => {
-  const router = useRouter();
+type CustomLinkProps = {
+  href: string;
+  title: string;
+  className?: string;
+};
+
+const CustomLink = ({ href, title, className = " " }: CustomLinkProps) => {
+  const pathname = usePathname();
   return (
     <Link href={href} className={`${className} relative group`}>
       {title}
@@ -21,7 +29,7 @@ const CustomLink = ({ href, title, className = " " }) => {
             h-[1px] inline-block bg-dark
             absolute left-0 -bottom-0.5
             group-hover:w-full transition-[width] ease duration-300
-            ${router.asPath === href ? "w-full" : "w-0"}
+            ${pathname === href ? "w-full" : "w-0"}
             `}
       >
         &nbsp;
