+++ src/components/Skills.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import React from 'react'
 import { motion } from 'framer-motion'
